CountBacteriaServer/requirements.txt
aniso8601==8.0.0
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
flask_cors==4.0.0
gunicorn==21.2.0
opencv-python==4.9.0.80
pillow==9.4.0
matplotlib==3.7.2
tqdm==4.65.0
torch==2.2.0
numpy==1.24.3